testing separate classes in same test file
